Replace old Job class with new chord variant
